# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fashion Xpress API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers/me:
    get:
      operationId: getMyCustomerProfile
      tags: [customers]
      responses:
        "200":
          description: Customer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateMyCustomerProfile
      tags: [customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/me/addresses:
    get:
      operationId: listMyAddresses
      tags: [customers]
      responses:
        "200":
          description: Addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      operationId: createMyAddress
      tags: [customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
  /customers/me/addresses/{id}:
    patch:
      operationId: updateMyAddress
      tags: [customers]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "404":
          description: Not found
    delete:
      operationId: deleteMyAddress
      tags: [customers]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /customers/me/measurements:
    get:
      operationId: getMyMeasurements
      tags: [customers]
      responses:
        "200":
          description: Measurements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Measurement"
        "404":
          description: Not set yet
    put:
      operationId: setMyMeasurements
      tags: [customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeasurementInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Measurement"

  # ---------------- CATALOG ----------------
...
  /executives:
    get:
      operationId: listExecutives
      tags: [executives]
      responses:
        "200":
          description: Executives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Executive"
    post:
      operationId: createExecutive
      tags: [executives]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecutiveInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Executive"
  /executives/me/visits:
    get:
      operationId: listMyVisits
      tags: [executives]
      summary: Executive - list bookings assigned to me
      parameters:
        - name: date
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Assigned bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"

  # ---------------- INVOICES ----------------
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Customer:
      type: object
      required: [id, userId, name, email, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        favoriteColors: { type: array, items: { type: string } }
        favoriteBrands: { type: array, items: { type: string } }
        lifetimeSpend: { type: number }
        createdAt: { type: string, format: date-time }
    CustomerUpdate:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        favoriteColors: { type: array, items: { type: string } }
        favoriteBrands: { type: array, items: { type: string } }
    Address:
      type: object
      required: [id, customerId, label, line1, city, state, pincode, isDefault]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        label: { type: string }
        line1: { type: string }
        line2: { type: ["string", "null"] }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        landmark: { type: ["string", "null"] }
        lat: { type: ["number", "null"] }
        lng: { type: ["number", "null"] }
        isDefault: { type: boolean }
    AddressInput:
      type: object
      required: [label, line1, city, state, pincode]
      properties:
        label: { type: string }
        line1: { type: string }
        line2: { type: string }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        landmark: { type: string }
        lat: { type: number }
        lng: { type: number }
        isDefault: { type: boolean }
    Measurement:
      type: object
      required: [id, customerId]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        gender: { type: ["string", "null"] }
        age: { type: ["integer", "null"] }
        heightCm: { type: ["number", "null"] }
        weightKg: { type: ["number", "null"] }
        topSize: { type: ["string", "null"] }
        bottomSize: { type: ["string", "null"] }
        preferredFit: { type: ["string", "null"] }
    MeasurementInput:
      type: object
      properties:
        gender: { type: string }
        age: { type: integer }
        heightCm: { type: number }
        weightKg: { type: number }
        topSize: { type: string }
        bottomSize: { type: string }
        preferredFit: { type: string }

    # ---------- CATALOG ----------
...
    Product:
      type: object
      required:
        [
          id,
          name,
          sku,
          categoryId,
          categoryName,
          brandId,
          brandName,
          color,
          sizes,
          fabric,
          occasion,
          mrp,
          sellingPrice,
          stock,
          images,
          rating,
          createdAt,
        ]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        sku: { type: string }
        barcode: { type: ["string", "null"] }
        qrCode: { type: ["string", "null"] }
        categoryId: { type: integer }
        categoryName: { type: string }
        brandId: { type: integer }
        brandName: { type: string }
        color: { type: string }
        sizes: { type: array, items: { type: string } }
        fabric: { type: string }
        occasion: { type: string }
        mrp: { type: number }
        sellingPrice: { type: number }
        purchaseCost: { type: ["number", "null"] }
        supplier: { type: ["string", "null"] }
        stock: { type: integer }
        warehouse: { type: ["string", "null"] }
        rack: { type: ["string", "null"] }
        images: { type: array, items: { type: string } }
        rating: { type: number }
        createdAt: { type: string, format: date-time }
...
    BookingProduct:
      type: object
      required: [id, bookingId, productId, product, status]
      properties:
        id: { type: integer }
        bookingId: { type: integer }
        productId: { type: integer }
        product:
          $ref: "#/components/schemas/Product"
        status:
          type: string
          enum: [reserved, sold, returned, damaged]
        priceAtSale: { type: ["number", "null"] }
        isRecommended: { type: boolean }
...
    Booking:
      type: object
      required:
        [
          id,
          bookingCode,
          customerId,
          name,
          phone,
          email,
          addressText,
          gender,
          preferredDate,
          preferredTime,
          status,
          products,
          createdAt,
        ]
      properties:
        id: { type: integer }
        bookingCode: { type: string }
        customerId: { type: integer }
        name: { type: string }
        phone: { type: string }
        email: { type: string }
        addressText: { type: string }
        lat: { type: ["number", "null"] }
        lng: { type: ["number", "null"] }
        gender: { type: string }
        age: { type: ["integer", "null"] }
        heightCm: { type: ["number", "null"] }
        weightKg: { type: ["number", "null"] }
        topSize: { type: ["string", "null"] }
        bottomSize: { type: ["string", "null"] }
        preferredFit: { type: ["string", "null"] }
        budget: { type: ["number", "null"] }
        occasion: { type: ["string", "null"] }
        preferredDate: { type: string }
        preferredTime: { type: string }
        preferredColors: { type: array, items: { type: string } }
        preferredBrands: { type: array, items: { type: string } }
        notes: { type: ["string", "null"] }
        status:
          type: string
          enum:
            [
              pending,
              confirmed,
              rescheduled,
              executive_assigned,
              in_progress,
              completed,
              cancelled,
              rejected,
            ]
        executiveId: { type: ["integer", "null"] }
        executiveName: { type: ["string", "null"] }
        products:
          type: array
          items:
            $ref: "#/components/schemas/BookingProduct"
        createdAt: { type: string, format: date-time }
...
    Executive:
      type: object
      required: [id, name, phone, email, activeBookings, rating, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        name: { type: string }
        phone: { type: string }
        email: { type: string }
        photoUrl: { type: ["string", "null"] }
        rating: { type: number }
        activeBookings: { type: integer }
        createdAt: { type: string, format: date-time }
    ExecutiveInput:
      type: object
      required: [name, phone, email, password]
      properties:
        name: { type: string }
        phone: { type: string }
        email: { type: string }
        password: { type: string, minLength: 6 }
        photoUrl: { type: string }

    # ---------- INVOICES ----------
...
    Notification:
      type: object
      required: [id, title, message, isRead, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        message: { type: string }
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }
...